remove unused files in new system
