+++ src/Makefile
@@ -2,7 +2,7 @@ PS2SDK=/usr/local/ps2dev/ps2sdk
 
 EE_BIN=test.elf
 #EE_OBJS=main.o gs.o mesh.o draw.o math.o pad.o drawstate.o script.o luadrawstate.o
-EE_OBJS=ps2lua.o gslua.o dmalua.o bufferlua.o tga.o padlua.o
+EE_OBJS=main.o gslua.o dmalua.o bufferlua.o tga.o padlua.o
 EE_LIBS=-ldma -lgraph -ldraw -lkernel -ldebug -lmath3d -lm -lpad -llua
 EE_CFLAGS+=-Wall --std=c99 -Wno-sign-compare -fno-strict-aliasing -fno-exceptions -DLUA_USE_PS2
 
@@ -17,7 +17,6 @@ endif
 .PHONY: clean
 clean:
 	rm -rf $(EE_BIN) $(EE_OBJS)
-	make -C lua -f makefile clean
