--- src/gs.h
@@ -1,19 +1,0 @@
-#ifndef GS_H
-#define GS_H
-
-#include <draw.h>
-#include <graph.h>
-
-struct draw_state {
-  int width;
-  int height;
-  int vmode;
-  int gmode;
-  framebuffer_t fb;
-  zbuffer_t zb;
-  char clear_col[3];
-};
-
-int gs_init(struct draw_state *ds, int psm, int psmz);
-
-#endif
